Slight refactor to kill the redundant layout page
