+++ lib/widgets/layout/header.dart
@@ -52,30 +52,31 @@ class Header extends StatelessWidget {
                     Material(
                         color: Colors.transparent,
                         child: SizedBox(
-                          height: 56,
-                          width: 56,
-                          child: IconButton(
-                              splashRadius: 56 / 2,
-                              splashColor: Colors.black.withOpacity(0.33),
-                              hoverColor: Colors.white.withOpacity(0.5),
-                              onPressed: () => onReset(context),
-                              icon: const Icon(
-                                Icons.restart_alt,
-                                color: Colors.black,
-                                size: 24,
-                              )
-                          ),
+                            height: 56,
+                            width: 56,
+                            child: IconButton(
+                                splashRadius: 56 / 2,
+                                splashColor: Colors.black.withOpacity(0.33),
+                                hoverColor: Colors.white.withOpacity(0.5),
+                                onPressed: () => onReset(context),
+                                icon: const Icon(
+                                    Icons.restart_alt,
+                                    color: Colors.black,
+                                    size: 24,
+                                )
+                            ),
                         )
                     )
                 ]
                 : [
                     Text(
                         isMini ? '${quiz.currentPlace + 1}/${quiz.questions.length}' : 'Question ${quiz.currentPlace + 1}/${quiz.questions.length}',
-                        style: const TextStyle(
-                            fontSize: 12,
+                        style: TextStyle(
+                            fontSize: isMobile ? 12 : 14,
                             fontWeight: FontWeight.w300,
                         ),
                     ),
+                    SizedBox(width: isMobile ? 12 : 24)
                 ];
 
         return SizedBox(
@@ -84,18 +85,29 @@ class Header extends StatelessWidget {
                 crossAxisAlignment: CrossAxisAlignment.center,
                 children: [
                     SizedBox(width: isMobile ? 12 : 16),
-                    const Image(
-                        image: AssetImage('assets/images/logo.png'),
-                        height: 40,
-                        filterQuality: FilterQuality.medium,
-                    ),
-                    SizedBox(width: isMini ? 8 : 12),
-                    Text(
-                        'LIBERTY COMPASS',
-                        style: TextStyle(
-                            fontSize: isMini ? 16 : 18,
-                            fontWeight: FontWeight.w300,
-                        )
+                    MouseRegion(
+                      cursor: SystemMouseCursors.click,
+                      child: GestureDetector(
+                        onTap: () => onReset(context),
+                        child: Row(
+                          crossAxisAlignment: CrossAxisAlignment.center,
+                          children: [
+                            const Image(
+                              image: AssetImage('assets/images/logo.png'),
+                              height: 40,
+                              filterQuality: FilterQuality.medium,
+                            ),
+                            SizedBox(width: isMini ? 8 : 12),
+                            Text(
+                                'LIBERTY COMPASS',
+                                style: TextStyle(
+                                  fontSize: isMini ? 16 : 18,
+                                  fontWeight: FontWeight.w300,
+                                )
+                            ),
+                          ],
+                        ),
+                      )
                     ),
                     Expanded(
                         child: Row(
